pyproject.toml
[project]
name = "agents-use-cases"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "crewai==0.102",
    "crewai-tools>=0.37.0",
    "databricks-vectorsearch",
    "databricks-agents>=0.21.2",
    "databricks-connect>=16.1.3",
    "databricks-langchain>=0.4.2",
    "databricks-sdk>=0.52.0",
    "faker>=37.1.0",
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langgraph==0.3.4",
    "mlflow>=2.22.0",
    "notebook>=7.4.1",
    "openai>=1.77.0",
    "openai-agents>=0.0.14",
    "pandas>=2.2.3",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "unitycatalog-ai",
    "unitycatalog-openai",
    "unitycatalog-crewai[databricks]>=0.2.0",
    "unitycatalog-openai[databricks]>=0.2.0",
    "tiktoken",
    "langchain-text-splitters",
    "openai-agents",
    "uv"
]

[tool.uv.sources]
unitycatalog-ai = { git = "https://github.com/unitycatalog/unitycatalog.git", subdirectory = "ai/core" }
unitycatalog-openai = { git = "https://github.com/unitycatalog/unitycatalog.git", subdirectory = "ai/integrations/openai" }
